+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Files show in uncommitted queue when automatically added (#407)
 - WebUI workspace view now works properly for filenames with spaces (#423)
 - Fixed error popups in interop editors in Studio on 2024.1 (#417)
+- Reintroduced amend (#425)
 
 ## [2.4.0] - 2024-07-08